+++ src/model/ColumnInfo.java
@@ -1,4 +1,4 @@
-package control;
+package model;
 
 import org.xml.sax.Attributes;
